+++ Manager/VboIndexer.h
@@ -25,7 +25,7 @@ namespace Manager {
                 std::vector<glm::vec2> & in_uvs,
                 std::vector<glm::vec3> & in_normals,
 
-                std::vector<unsigned short> & out_indices,
+                std::vector<unsigned int> & out_indices,
                 std::vector<glm::vec3> & out_vertices,
                 std::vector<glm::vec2> & out_uvs,
                 std::vector<glm::vec3> & out_normals
@@ -37,7 +37,7 @@ namespace Manager {
                 std::vector<glm::vec3> & in_tangents,
                 std::vector<glm::vec3> & in_bitangents,
 
-                std::vector<unsigned short> & out_indices,
+                std::vector<unsigned int> & out_indices,
                 std::vector<glm::vec3> & out_vertices,
                 std::vector<glm::vec2> & out_uvs,
                 std::vector<glm::vec3> & out_normals,
